move values between columns to suit values in 2020 + bettter spacing
Moved a few values between columns so that viewed in year 2020 figures,
they're mostly rising.  In trying to remember the values, I'll likely remember
what column & thus what order of magnitude they're in.

The only exception is reading 1MiB from disk as it makes better sense in
context with the disk seek in the same column.

- Also stopped the bleed of the 'box' column into the legend